bash-itunes shuffle function fix for iTunes 11, this fix is bringing back the feature of showing the current status of shuffle and turning on and off shuffle for iTunes 11
With iTunes 11 it wasn't possible to use the shuffle commands properly within bash-itunes. This patch fixes that. Instead of calling _itunes_tell in the _cmd_shuffle section, everything will be called via _osascript, since iTunes won't be called directly but rather indirectly via the iTunes process.
The shuffle function is now realized via the System UI and therefore it could be running against problems, if the system language is not "English". This was not tested.